+++ lib/src/elements/th_line.dart
@@ -136,9 +136,15 @@ class THLine extends THElement
           originalLineInTH2File ?? this.originalLineInTH2File,
       lineType: lineType ?? this.lineType,
       unknownPLAType: unknownPLAType ?? this.unknownPLAType,
-      childrenMPIDs: childrenMPIDs ?? this.childrenMPIDs,
-      optionsMap: optionsMap ?? this.optionsMap,
-      attrOptionsMap: attrOptionsMap ?? this.attrOptionsMap,
+      childrenMPIDs: childrenMPIDs ?? this.childrenMPIDs.toList(),
+      optionsMap:
+          optionsMap ??
+          SplayTreeMap<THCommandOptionType, THCommandOption>.from(
+            this.optionsMap,
+          ),
+      attrOptionsMap:
+          attrOptionsMap ??
+          SplayTreeMap<String, THAttrCommandOption>.from(this.attrOptionsMap),
     );
   }